refactor: sane module hierarchy, remove reloading
