+++ src/commands/lua_command.rs
@@ -1,28 +1,30 @@
-use std::sync::Arc;
-
-use serenity::all::{CommandDataOptionValue, CommandInteraction, Http};
-use tokio::sync::Mutex;
-
-use crate::{
-    commands::{execute::extensions::TemporaryChannelUpdate, lua_registry::CommandRegistry},
-    config,
+use std::{
+    collections::HashMap,
+    sync::{Arc, Mutex},
 };
 
-use super::lua_executor::execute_lua_thread;
+use serenity::all::{
+    CommandDataOptionValue, CommandInteraction, CommandOptionType, CreateCommand,
+    CreateCommandOption, Http,
+};
+
+use crate::{
+    config,
+    lua::{execute_lua_thread, extensions::TemporaryChannelUpdate},
+};
 
 pub struct Handler {
     name: String,
     discord_config: config::Discord,
-    command_registry: CommandRegistry,
-    global_lua: Arc<Mutex<mlua::Lua>>,
+    command_registry: LuaCommandRegistry,
+    global_lua: mlua::Lua,
 }
-
 impl Handler {
     pub fn new(
         name: String,
         discord_config: config::Discord,
-        command_registry: CommandRegistry,
-        global_lua: Arc<Mutex<mlua::Lua>>,
+        command_registry: LuaCommandRegistry,
+        global_lua: mlua::Lua,
     ) -> Self {
         Self {
             name,
@@ -32,7 +34,6 @@ impl Handler {
         }
     }
 }
-
 #[serenity::async_trait]
 impl super::CommandHandler for Handler {
     async fn register(&self, http: &Http) -> anyhow::Result<()> {
@@ -56,10 +57,11 @@ impl super::CommandHandler for Handler {
         let (print_tx, print_rx) = flume::unbounded::<String>();
 
         // Lock the global Lua state for this execution (held for entire duration)
-        let lua = self.global_lua.lock().await;
+        let lua = &self.global_lua;
 
         // Update output channels for this execution
-        let _temporary_channel_update = TemporaryChannelUpdate::new(&lua, output_tx, print_tx)?;
+        let _temporary_channel_update =
+            TemporaryChannelUpdate::new(lua.clone(), output_tx, print_tx)?;
 
         // Build interaction table
         let interaction = lua.create_table()?;
@@ -113,5 +115,63 @@ impl super::CommandHandler for Handler {
             None,
         )
         .await
+    }
+}
+
+pub type LuaCommandRegistry = Arc<Mutex<HashMap<String, LuaCommand>>>;
+
+// LuaCommand stores command metadata and handler function reference
+pub struct LuaCommand {
+    pub name: String,
+    pub description: String,
+    pub options: Vec<LuaCommandOption>,
+    pub handler: mlua::Function,
+}
+#[derive(Clone)]
+pub struct LuaCommandOption {
+    pub name: String,
+    pub description: String,
+    pub option_type: CommandOptionType,
+    pub required: bool,
+    pub min_value: Option<f64>,
+    pub max_value: Option<f64>,
+    pub min_length: Option<u16>,
+    pub max_length: Option<u16>,
+    pub autocomplete: bool,
+    pub choices: Vec<(String, String)>, // (name, value) for string choices
+}
+impl LuaCommand {
+    pub fn to_discord_command(&self) -> CreateCommand {
+        let mut cmd = CreateCommand::new(&self.name).description(&self.description);
+
+        for opt in &self.options {
+            let mut option = CreateCommandOption::new(opt.option_type, &opt.name, &opt.description)
+                .required(opt.required);
+
+            if let Some(min_value) = opt.min_value {
+                option = option.min_number_value(min_value);
+            }
+            if let Some(max_value) = opt.max_value {
+                option = option.max_number_value(max_value);
+            }
+            if let Some(min_length) = opt.min_length {
+                option = option.min_length(min_length);
+            }
+            if let Some(max_length) = opt.max_length {
+                option = option.max_length(max_length);
+            }
+            if opt.autocomplete {
+                option = option.set_autocomplete(true);
+            }
+
+            // Add choices if present
+            for (choice_name, choice_value) in &opt.choices {
+                option = option.add_string_choice(choice_name, choice_value);
+            }
+
+            cmd = cmd.add_option(option);
+        }
+
+        cmd
     }
 }
